Core refactoring and condoor renaming
